✨ added new sharepic canvas component for calculating the colorset by the selected image
