+++ src/components/templates/EditorTextOnly.tsx
@@ -9,6 +9,7 @@ import { EditorText } from "../organism/EditorText";
 import { EditorPictogram } from "../organism/EditorPictogram";
 import { EditorPictogramPosition } from "../organism/EditorPictogramPosition";
 import { pictograms } from "../../constants/pictograms";
+import { EditorBackgroundImage } from "../organism/EditorBackgroundImage";
 
 export default function EditorTextOnly() {
 	const { state, dispatch } = useSharePic();
@@ -21,6 +22,17 @@ export default function EditorTextOnly() {
 				handleColorSet={(e: { target: { value: string; }; }) => dispatch({
 					type: "SET_COLOR_SET",
 					payload: colorSets.find((c) => c.name === e.target.value)!!,
+				})}
+			/>
+
+			<EditorBackgroundImage
+				handleBackgroundImage={(base64File: string) => dispatch({
+					type: "SET_BACKGROUND_IMAGE",
+					payload: base64File
+				})}
+				handleFileDelete={() => dispatch({
+					type: "SET_BACKGROUND_IMAGE",
+					payload: null
 				})}
 			/>
