+++ src/components/templates/SharePicPictogramOnly.tsx
@@ -5,6 +5,7 @@ import { SharePicArrow } from "../organism/SharePicArrow";
 import { SharePicPictogram } from "../organism/SharePicPictogram";
 import { colorSets, highlightColors } from "../../constants/colors";
 import { pictograms } from "../../constants/pictograms";
+import { SharePicCanvas } from "../organism/SharePicCanvas";
 
 export default function SharePicTextOnly() {
 
@@ -30,7 +31,14 @@ export default function SharePicTextOnly() {
 	}, []);
 
 	return(
-		<div className='sharepic-canvas' style={{ background: state.colorSet.backgroundColor }}>
+		<SharePicCanvas
+			backgroundImage={state.backgroundImage}
+			colorSet={state.colorSet}
+			handleColorSet={(colorSet: ColorSet) => dispatch({
+				type: "SET_COLOR_SET",
+				payload: colorSet
+			})}
+		>
 			<SharePicLogo
 				localGroup={state.localGroup}
 				visible={state.logoVisible}
@@ -45,6 +53,6 @@ export default function SharePicTextOnly() {
 				position={state.pictogramPosition}
 				colorSet={state.colorSet}
 			/>
-		</div>
+		</SharePicCanvas>
 	);
 }
