+++ src/components/templates/SharePicTitleAndText.tsx
@@ -7,6 +7,7 @@ import { SharePicText } from "../organism/SharePicText";
 import { SharePicPictogram } from "../organism/SharePicPictogram";
 import { colorSets, highlightColors } from "../../constants/colors";
 import { SharePicContent } from "../organism/SharePicContent";
+import { SharePicCanvas } from "../organism/SharePicCanvas";
 
 export default function SharePicTitleAndText() {
 
@@ -32,7 +33,14 @@ export default function SharePicTitleAndText() {
 	}, []);
 
 	return(
-		<div className='sharepic-canvas' style={{ background: state.colorSet.backgroundColor }}>
+		<SharePicCanvas
+			backgroundImage={state.backgroundImage}
+			colorSet={state.colorSet}
+			handleColorSet={(colorSet: ColorSet) => dispatch({
+				type: "SET_COLOR_SET",
+				payload: colorSet
+			})}
+		>
 			<SharePicLogo
 				localGroup={state.localGroup}
 				visible={state.logoVisible}
@@ -63,6 +71,6 @@ export default function SharePicTitleAndText() {
 				position={state.pictogramPosition}
 				colorSet={state.colorSet}
 			/>
-		</div>
+		</SharePicCanvas>
 	);
 }
