+++ src/components/templates/SharePicTitleOnly.tsx
@@ -7,6 +7,7 @@ import { SharePicHeading } from "../organism/SharePicHeading";
 import { SharePicPictogram } from "../organism/SharePicPictogram";
 import { colorSets, highlightColors } from "../../constants/colors";
 import { pictograms } from "../../constants/pictograms";
+import { SharePicCanvas } from "../organism/SharePicCanvas";
 
 export default function SharePicTitleOnly() {
 
@@ -31,82 +32,16 @@ export default function SharePicTitleOnly() {
 		});
 	}, []);
 
-	useEffect(() => {
-		const calculateAndSetColorSet = async () => {
-			if (state.backgroundImage) {
-				try {
-					const colorSet = await calculateColorSetOnImageBrightness(state.backgroundImage);
-					dispatch({ type: "SET_COLOR_SET", payload: colorSet });
-				} catch (error) {
-					console.error("Fehler beim Berechnen des Farbsatzes:", error);
-				}
-			}
-		};
-
-		calculateAndSetColorSet();
-	}, [state.backgroundImage, dispatch]);
-
-	const calculateColorSetOnImageBrightness = (base64Image: string): Promise<ColorSet> => {
-		const calculateImageBrightness = (image: HTMLImageElement): number => {
-			const canvas = document.createElement("canvas");
-			const ctx = canvas.getContext("2d");
-
-			if (!ctx) {
-				throw new Error("Canvas context not available");
-			}
-
-			canvas.width = image.width;
-			canvas.height = image.height;
-
-			ctx.drawImage(image, 0, 0, image.width, image.height);
-
-			const imageData = ctx.getImageData(0, 0, image.width, image.height);
-			const pixels = imageData.data;
-
-			let totalBrightness = 0;
-			let pixelCount = 0;
-
-			for (let i = 0; i < pixels.length; i += 4) {
-				const r = pixels[i];
-				const g = pixels[i + 1];
-				const b = pixels[i + 2];
-
-				const brightness = 0.2126 * r + 0.7152 * g + 0.0722 * b;
-				totalBrightness += brightness;
-				pixelCount++;
-			}
-
-			return totalBrightness / pixelCount;
-		};
-
-		return new Promise((resolve, reject) => {
-			const image = new Image();
-			image.src = base64Image;
-			console.log(base64Image);
-
-
-			image.onload = () => {
-				const brightness = calculateImageBrightness(image);
-
-				if (brightness < 180) {
-					resolve(colorSets.find(cS => cS.name === "Black") || colorSets[6]);
-				} else {
-					resolve(colorSets.find(cS => cS.name === "White") || colorSets[5]);
-				}
-			};
-
-			image.onerror = () => {
-			reject(new Error("Failed to load base64 image"));
-			};
-		});
-	};
 
 	return(
-		<div className='sharepic-canvas' style={{
-				background: state.backgroundImage ? `url(${state.backgroundImage})` : state.colorSet.backgroundColor,
-				backgroundSize: state.backgroundImage ? `cover` : "none",
-				backgroundPosition: state.backgroundImage ? `cover` : "none"
-		}}>
+		<SharePicCanvas
+			backgroundImage={state.backgroundImage}
+			colorSet={state.colorSet}
+			handleColorSet={(colorSet: ColorSet) => dispatch({
+				type: "SET_COLOR_SET",
+				payload: colorSet
+			})}
+		>
 			<SharePicLogo
 				localGroup={state.localGroup}
 				visible={state.logoVisible}
@@ -132,6 +67,6 @@ export default function SharePicTitleOnly() {
 				position={state.pictogramPosition}
 				colorSet={state.colorSet}
 			/>
-		</div>
+		</SharePicCanvas>
 	);
 }
